Create Assignment_3
Summary
Using the Object-Oriented UML model provided, create an Object-Oriented program that creates a simple banking application which allows the user to open an account, select an account, withdraw and deposit money, check balance. Analyze the partial class diagram to determine the structure of the program and the program flow. Practice program flow control statements to branch and repeat statements as needed. Practice the principles of separation of concerns to divide the business logic appropriately between classes and to create additional methods as needed